requirements.txt
torch
torchvision
decord
open-clip-torch
tqdm
numpy